refactor(ui): share one useClickOutside hook
